+++ commands/System/ping.js
@@ -5,7 +5,7 @@ exports.run = async (client, msg) => {
 
 exports.conf = {
   enabled: true,
-  guildOnly: false,
+  runIn: ["text", "dm", "group"],
   aliases: [],
   permLevel: 0,
   botPerms: [],