+++ commands/System/stats.js
@@ -18,7 +18,7 @@ exports.run = (client, msg) => {
 
 exports.conf = {
   enabled: true,
-  guildOnly: false,
+  runIn: ["text", "dm", "group"],
   aliases: ["details", "what"],
   permLevel: 0,
   botPerms: [],